@@ -6,7 +6,7 @@ var config = {
   port: process.env.PORT || 8889,
   karmaPort: 9876,
   ssr: true,
-
+  sentry: false,
   // This part goes to React-Helmet for Head of our HTML
   app: {
     head: {
