@@ -1,5 +1,8 @@
 import {History} from "history";
 const appConfig = require("../../../config/main");
+const localConfig = require("../../../config/main.local");
+import * as createRavenMiddleware from "raven-for-redux";
+import * as Raven from "raven-js";
 import {routerMiddleware} from "react-router-redux";
 import {applyMiddleware, compose, createStore} from "redux";
 import {createLogger} from "redux-logger";
@@ -14,6 +17,7 @@ interface IExtendedStore extends Redux.Store<IStore> {
 
 export function configureStore(history: History, initialState?: IStore): IExtendedStore {
 
+  const mergedConfig = {...appConfig, ...localConfig};
   const sagaMiddleware = createSagaMiddleware();
   const middlewares: Redux.Middleware[] = [
     routerMiddleware(history),
@@ -21,12 +25,17 @@ export function configureStore(history: History, initialState?: IStore): IExtend
   ];
 
   /** Add Only Dev. Middlewares */
-  if (appConfig.env !== "production" && process.env.BROWSER) {
+  if (mergedConfig.env !== "production" && process.env.BROWSER) {
     const logger = createLogger();
     middlewares.push(logger);
   }
 
-  const composeEnhancers = (appConfig.env !== "production" &&
+  if (mergedConfig.sentry && process.env.BROWSER) {
+    Raven.config(mergedConfig.sentry.dsn).setRelease(mergedConfig.sentry.release).install();
+    middlewares.unshift(createRavenMiddleware(Raven));
+  }
+
+  const composeEnhancers = (mergedConfig.env !== "production" &&
     typeof window === "object" &&
     (typeof window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ === "function") &&
     window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__({shouldHotReload: false})) || compose;
@@ -35,7 +44,7 @@ export function configureStore(history: History, initialState?: IStore): IExtend
     applyMiddleware(...middlewares)
   ));
 
-  if (appConfig.env === "development" && (module as any).hot) {
+  if (mergedConfig.env === "development" && (module as any).hot) {
     (module as any).hot.accept("./rootReducer", () => {
       store.replaceReducer((require("./rootReducer").default));
     });
