@@ -1,5 +1,5 @@
-import {IStars, IStarsAction} from "models/starsModel";
 import promiseReducer, {IRequestType} from "../../helpers/promiseReducer";
+import {IAction, IState} from "./baseModule";
 
 export const LOAD_STARS: string = "stars/LOAD_STARS";
 export const requestType: IRequestType = {
@@ -8,16 +8,27 @@ export const requestType: IRequestType = {
   SUCCESS: "stars/SUCCESS"
 };
 
-/** Initial State */
-const initialState: IStars = {
-  payload: null
+export interface IStars {
+  stargazers_count: number;
+}
+
+const initialState: IState<IStars> = {
+  isFetching: true,
+  payload: {
+    stargazers_count: -1
+  }
 };
 
-/** Reducer */
-export function starsReducer(state: IStars = initialState, action: IStarsAction): IStars {
+export function starsReducer(state: IState<IStars> = initialState, action: IAction<IStars>): IState<IStars> {
   // an example how you can deal with action type other than requesting types in promiseReducer
   if (action.type === LOAD_STARS) {
     return state;
   }
-  return promiseReducer<IStars, IStarsAction>(requestType, state, action);
+  return promiseReducer<IStars>(requestType, state, action);
+}
+
+export function loadStars(): IAction<IStars> {
+  return {
+    type: LOAD_STARS
+  };
 }
